add-retrofit+models
all logic is ready.
todo: debug why it's not returning anything.

+++ app/src/main/java/com/example/shopify/core/common/data/model/DraftOrderRegistration.kt
@@ -5,6 +5,6 @@ import com.google.gson.annotations.SerializedName
 data class DraftOrderRegistration(
     @SerializedName("line_items") val lineItems: List<LineItemRegistration>,
     @SerializedName("applied_discount") val appliedDiscount: DiscountRegistration,
-    val customer: Int,
+    val customer: Long,
     @SerializedName("use_customer_default_address") val useCustomerDefaultAddress: Boolean
 )
